+++ bigframes/core/nodes.py
@@ -15,7 +15,7 @@
 from __future__ import annotations
 
 import abc
-from dataclasses import dataclass, field, fields, replace
+import dataclasses
 import datetime
 import functools
 import itertools
@@ -46,13 +46,13 @@ OVERHEAD_VARIABLES = 5
 COLUMN_SET = frozenset[bfet_ids.ColumnId]
 
 
-@dataclass(frozen=True)
+@dataclasses.dataclass(frozen=True)
 class Field:
     id: bfet_ids.ColumnId
     dtype: bigframes.dtypes.Dtype
 
 
-@dataclass(eq=False, frozen=True)
+@dataclasses.dataclass(eq=False, frozen=True)
 class BigFrameNode(abc.ABC):
     """
     Immutable node for representing 2D typed array as a tree of operators.
@@ -131,7 +131,7 @@ class BigFrameNode(abc.ABC):
 
     def _as_tuple(self) -> Tuple:
         """Get all fields as tuple."""
-        return tuple(getattr(self, field.name) for field in fields(self))
+        return tuple(getattr(self, field.name) for field in dataclasses.fields(self))
 
     def __hash__(self) -> int:
         # Custom hash that uses cache to avoid costly recomputation
@@ -286,7 +286,7 @@ class BigFrameNode(abc.ABC):
         return self.transform_children(lambda x: x.prune(used_cols))
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class UnaryNode(BigFrameNode):
     child: BigFrameNode
 
@@ -305,18 +305,14 @@ class UnaryNode(BigFrameNode):
     def transform_children(
         self, t: Callable[[BigFrameNode], BigFrameNode]
     ) -> BigFrameNode:
-        transformed = replace(self, child=t(self.child))
+        transformed = dataclasses.replace(self, child=t(self.child))
         if self == transformed:
             # reusing existing object speeds up eq, and saves a small amount of memory
             return self
         return transformed
 
-    def replace_child(
-        self, new_child: BigFrameNode, validate: bool = False
-    ) -> UnaryNode:
-        new_self = replace(self, child=new_child)  # type: ignore
-        if validate:
-            new_self._validate()
+    def replace_child(self, new_child: BigFrameNode) -> UnaryNode:
+        new_self = dataclasses.replace(self, child=new_child)  # type: ignore
         return new_self
 
     @property
@@ -324,7 +320,7 @@ class UnaryNode(BigFrameNode):
         return self.child.order_ambiguous
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class SliceNode(UnaryNode):
     """Logical slice node conditionally becomes limit or filter over row numbers."""
 
@@ -387,7 +383,7 @@ class SliceNode(UnaryNode):
         return self
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class JoinNode(BigFrameNode):
     left_child: BigFrameNode
     right_child: BigFrameNode
@@ -449,7 +445,7 @@ class JoinNode(BigFrameNode):
     def transform_children(
         self, t: Callable[[BigFrameNode], BigFrameNode]
     ) -> BigFrameNode:
-        transformed = replace(
+        transformed = dataclasses.replace(
             self, left_child=t(self.left_child), right_child=t(self.right_child)
         )
         if self == transformed:
@@ -479,10 +475,10 @@ class JoinNode(BigFrameNode):
             )
             for l_cond, r_cond in self.conditions
         )
-        return replace(self, conditions=new_conds)  # type: ignore
+        return dataclasses.replace(self, conditions=new_conds)  # type: ignore
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class ConcatNode(BigFrameNode):
     # TODO: Explcitly map column ids from each child
     children: Tuple[BigFrameNode, ...]
@@ -538,7 +534,9 @@ class ConcatNode(BigFrameNode):
     def transform_children(
         self, t: Callable[[BigFrameNode], BigFrameNode]
     ) -> BigFrameNode:
-        transformed = replace(self, children=tuple(t(child) for child in self.children))
+        transformed = dataclasses.replace(
+            self, children=tuple(t(child) for child in self.children)
+        )
         if self == transformed:
             # reusing existing object speeds up eq, and saves a small amount of memory
             return self
@@ -552,13 +550,13 @@ class ConcatNode(BigFrameNode):
         self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]
     ) -> BigFrameNode:
         new_ids = tuple(mappings.get(id, id) for id in self.output_ids)
-        return replace(self, output_ids=new_ids)
+        return dataclasses.replace(self, output_ids=new_ids)
 
     def remap_refs(self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]):
         return self
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class FromRangeNode(BigFrameNode):
     # TODO: Enforce single-row, single column constraint
     start: BigFrameNode
@@ -606,7 +604,7 @@ class FromRangeNode(BigFrameNode):
     def transform_children(
         self, t: Callable[[BigFrameNode], BigFrameNode]
     ) -> BigFrameNode:
-        transformed = replace(self, start=t(self.start), end=t(self.end))
+        transformed = dataclasses.replace(self, start=t(self.start), end=t(self.end))
         if self == transformed:
             # reusing existing object speeds up eq, and saves a small amount of memory
             return self
@@ -619,7 +617,9 @@ class FromRangeNode(BigFrameNode):
     def remap_vars(
         self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]
     ) -> BigFrameNode:
-        return replace(self, output_id=mappings.get(self.output_id, self.output_id))
+        return dataclasses.replace(
+            self, output_id=mappings.get(self.output_id, self.output_id)
+        )
 
     def remap_refs(self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]):
         return self
@@ -628,7 +628,7 @@ class FromRangeNode(BigFrameNode):
 # Input Nodex
 # TODO: Most leaf nodes produce fixed column names based on the datasource
 # They should support renaming
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class LeafNode(BigFrameNode):
     @property
     def roots(self) -> typing.Set[BigFrameNode]:
@@ -654,12 +654,12 @@ class ScanItem(typing.NamedTuple):
     source_id: str  # Flexible enough for both local data and bq data
 
 
-@dataclass(frozen=True)
+@dataclasses.dataclass(frozen=True)
 class ScanList:
     items: typing.Tuple[ScanItem, ...]
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class ReadLocalNode(LeafNode):
     feather_bytes: bytes
     data_schema: schemata.ArraySchema
@@ -725,20 +725,20 @@ class ReadLocalNode(LeafNode):
                 for item in self.scan_list.items
             )
         )
-        return replace(self, scan_list=new_scan_list)
+        return dataclasses.replace(self, scan_list=new_scan_list)
 
     def remap_refs(self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]):
         return self
 
 
-@dataclass(frozen=True)
+@dataclasses.dataclass(frozen=True)
 class GbqTable:
-    project_id: str = field()
-    dataset_id: str = field()
-    table_id: str = field()
-    physical_schema: Tuple[bq.SchemaField, ...] = field()
-    n_rows: int = field()
-    is_physically_stored: bool = field()
+    project_id: str = dataclasses.field()
+    dataset_id: str = dataclasses.field()
+    table_id: str = dataclasses.field()
+    physical_schema: Tuple[bq.SchemaField, ...] = dataclasses.field()
+    n_rows: int = dataclasses.field()
+    is_physically_stored: bool = dataclasses.field()
     cluster_cols: typing.Optional[Tuple[str, ...]]
 
     @staticmethod
@@ -761,7 +761,7 @@ class GbqTable:
         )
 
 
-@dataclass(frozen=True)
+@dataclasses.dataclass(frozen=True)
 class BigqueryDataSource:
     """
     Google BigQuery Data source.
@@ -777,14 +777,14 @@ class BigqueryDataSource:
 
 
 ## Put ordering in here or just add order_by node above?
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class ReadTableNode(LeafNode):
     source: BigqueryDataSource
     # Subset of physical schema column
     # Mapping of table schema ids to bfet id.
     scan_list: ScanList
 
-    table_session: bigframes.session.Session = field()
+    table_session: bigframes.session.Session = dataclasses.field()
 
     def _validate(self):
         # enforce invariants
@@ -860,7 +860,7 @@ class ReadTableNode(LeafNode):
             tuple(item for item in self.scan_list.items if item.id in used_cols)
             or (self.scan_list.items[0],)
         )
-        return replace(self, scan_list=new_scan_list)
+        return dataclasses.replace(self, scan_list=new_scan_list)
 
     def remap_vars(
         self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]
@@ -871,21 +871,21 @@ class ReadTableNode(LeafNode):
                 for item in self.scan_list.items
             )
         )
-        return replace(self, scan_list=new_scan_list)
+        return dataclasses.replace(self, scan_list=new_scan_list)
 
     def remap_refs(self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]):
         return self
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class CachedTableNode(ReadTableNode):
     # The original BFET subtree that was cached
     # note: this isn't a "child" node.
-    original_node: BigFrameNode = field()
+    original_node: BigFrameNode = dataclasses.field()
 
 
 # Unary nodes
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class PromoteOffsetsNode(UnaryNode):
     col_id: bigframes.core.identifiers.ColumnId
 
@@ -933,13 +933,13 @@ class PromoteOffsetsNode(UnaryNode):
     def remap_vars(
         self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]
     ) -> BigFrameNode:
-        return replace(self, col_id=mappings.get(self.col_id, self.col_id))
+        return dataclasses.replace(self, col_id=mappings.get(self.col_id, self.col_id))
 
     def remap_refs(self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]):
         return self
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class FilterNode(UnaryNode):
     predicate: ex.Expression
 
@@ -970,7 +970,7 @@ class FilterNode(UnaryNode):
         return self
 
     def remap_refs(self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]):
-        return replace(
+        return dataclasses.replace(
             self,
             predicate=self.predicate.remap_column_refs(
                 mappings, allow_partial_bindings=True
@@ -978,7 +978,7 @@ class FilterNode(UnaryNode):
         )
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class OrderByNode(UnaryNode):
     by: Tuple[OrderingExpression, ...]
 
@@ -1028,10 +1028,10 @@ class OrderByNode(UnaryNode):
                 for by_expr in self.by
             ),
         )
-        return replace(self, by=new_by)
+        return dataclasses.replace(self, by=new_by)
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class ReversedNode(UnaryNode):
     # useless field to make sure has distinct hash
     reversed: bool = True
@@ -1062,7 +1062,7 @@ class ReversedNode(UnaryNode):
         return self
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class SelectionNode(UnaryNode):
     input_output_pairs: typing.Tuple[
         typing.Tuple[ex.DerefOp, bigframes.core.identifiers.ColumnId], ...
@@ -1122,17 +1122,17 @@ class SelectionNode(UnaryNode):
         new_pairs = tuple(
             (ref, mappings.get(id, id)) for ref, id in self.input_output_pairs
         )
-        return replace(self, input_output_pairs=new_pairs)
+        return dataclasses.replace(self, input_output_pairs=new_pairs)
 
     def remap_refs(self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]):
         new_fields = tuple(
             (ex.remap_column_refs(mappings, allow_partial_bindings=True), id)
             for ex, id in self.input_output_pairs
         )
-        return replace(self, input_output_pairs=new_fields)  # type: ignore
+        return dataclasses.replace(self, input_output_pairs=new_fields)  # type: ignore
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class ProjectionNode(UnaryNode):
     """Assigns new variables (without modifying existing ones)"""
 
@@ -1192,19 +1192,19 @@ class ProjectionNode(UnaryNode):
         self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]
     ) -> BigFrameNode:
         new_fields = tuple((ex, mappings.get(id, id)) for ex, id in self.assignments)
-        return replace(self, assignments=new_fields)
+        return dataclasses.replace(self, assignments=new_fields)
 
     def remap_refs(self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]):
         new_fields = tuple(
             (ex.remap_column_refs(mappings, allow_partial_bindings=True), id)
             for ex, id in self.assignments
         )
-        return replace(self, assignments=new_fields)
+        return dataclasses.replace(self, assignments=new_fields)
 
 
 # TODO: Merge RowCount into Aggregate Node?
 # Row count can be compute from table metadata sometimes, so it is a bit special.
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class RowCountNode(UnaryNode):
     col_id: bfet_ids.ColumnId = bfet_ids.ColumnId("count")
 
@@ -1239,7 +1239,7 @@ class RowCountNode(UnaryNode):
     def remap_vars(
         self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]
     ) -> BigFrameNode:
-        return replace(self, col_id=mappings.get(self.col_id, self.col_id))
+        return dataclasses.replace(self, col_id=mappings.get(self.col_id, self.col_id))
 
     def remap_refs(self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]):
         return self
@@ -1249,7 +1249,7 @@ class RowCountNode(UnaryNode):
         return self
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class AggregateNode(UnaryNode):
     aggregations: typing.Tuple[
         typing.Tuple[ex.Aggregation, bigframes.core.identifiers.ColumnId], ...
@@ -1320,7 +1320,7 @@ class AggregateNode(UnaryNode):
         self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]
     ) -> BigFrameNode:
         new_aggs = tuple((agg, mappings.get(id, id)) for agg, id in self.aggregations)
-        return replace(self, aggregations=new_aggs)
+        return dataclasses.replace(self, aggregations=new_aggs)
 
     def remap_refs(self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]):
         new_aggs = tuple(
@@ -1328,10 +1328,12 @@ class AggregateNode(UnaryNode):
             for agg, id in self.aggregations
         )
         new_by_ids = tuple(id.remap_column_refs(mappings) for id in self.by_column_ids)
-        return replace(self, by_column_ids=new_by_ids, aggregations=new_aggs)
+        return dataclasses.replace(
+            self, by_column_ids=new_by_ids, aggregations=new_aggs
+        )
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class WindowOpNode(UnaryNode):
     column_name: ex.DerefOp
     op: agg_ops.UnaryWindowOp
@@ -1392,12 +1394,12 @@ class WindowOpNode(UnaryNode):
     def remap_vars(
         self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]
     ) -> BigFrameNode:
-        return replace(
+        return dataclasses.replace(
             self, output_name=mappings.get(self.output_name, self.output_name)
         )
 
     def remap_refs(self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]):
-        return replace(
+        return dataclasses.replace(
             self,
             column_name=self.column_name.remap_column_refs(
                 mappings, allow_partial_bindings=True
@@ -1408,7 +1410,7 @@ class WindowOpNode(UnaryNode):
         )
 
 
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class RandomSampleNode(UnaryNode):
     fraction: float
 
@@ -1444,7 +1446,7 @@ class RandomSampleNode(UnaryNode):
 
 
 # TODO: Explode should create a new column instead of overriding the existing one
-@dataclass(frozen=True, eq=False)
+@dataclasses.dataclass(frozen=True, eq=False)
 class ExplodeNode(UnaryNode):
     column_ids: typing.Tuple[ex.DerefOp, ...]
 
@@ -1496,4 +1498,4 @@ class ExplodeNode(UnaryNode):
         self, mappings: Mapping[bfet_ids.ColumnId, bfet_ids.ColumnId]
     ) -> BigFrameNode:
         new_ids = tuple(id.remap_column_refs(mappings) for id in self.column_ids)
-        return replace(self, column_ids=new_ids)  # type: ignore
+        return dataclasses.replace(self, column_ids=new_ids)  # type: ignore
