+++ bigframes/core/rewrite/__init__.py
@@ -1,0 +1,26 @@
+# Copyright 2024 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from bigframes.core.rewrite.identifiers import remap_variables
+from bigframes.core.rewrite.implicit_align import try_join_as_projection
+from bigframes.core.rewrite.legacy_align import legacy_join_as_projection
+from bigframes.core.rewrite.slices import pullup_limit_from_slice, replace_slice_ops
+
+__all__ = [
+    "legacy_join_as_projection",
+    "try_join_as_projection",
+    "replace_slice_ops",
+    "pullup_limit_from_slice",
+    "remap_variables",
+]
